p2p: add a flag to only accept connection from a trusted peer (#43)
This commit adds flag --trusted-node-only, when using this flag, node will only
accept connection from a trusted peer. This is suboptimal as it does not change
the dial logic which still tries to dial not trusted peer. But it works for
now, the dial part can be in the future improvement.